requirements.txt
asgiref==3.8.1
Django==5.2
sqlparse==0.5.3
typing_extensions==4.13.2
anyio==4.9.0
certifi==2025.1.31
charset-normalizer==3.4.1
cloudinary==1.43.0
django-cloudinary-storage==0.3.0
django-cors-headers==4.7.0
django_paystack==1.0.0
djangorestframework==3.16.0
djangorestframework_simplejwt==5.5.0
exceptiongroup==1.2.2
h11==0.14.0
httpcore==1.0.7
httpx==0.28.1
idna==3.10
pillow==11.1.0
PyJWT==2.9.0
pypaystack2==2.0.3
requests==2.32.3
six==1.17.0
sniffio==1.3.1
urllib3==2.3.0
